Update ExportController.php
This line of code will fix this bug I've found in my applications:

When I was exporting a GridView all cell values starting with a zero (0) where cutted of, so the zero was gone. By adding a zero width non-joiner (zwnj) in front of the cell values the problem is gone. This fix is only for excel files, as they are the only file type that has problems with leading zeros.

I don't know if this is really a fix for something, just wanted to let you know and offer a solution.