Improved the leading zero handling
I've seen some issues with the old work around. For example: If there was a date field "09.12.2015" it would be replaced by "&zwnj;09.12.2015" this would cause that the date column would be not sortable anymore.

So now the Controller is looking for only <td></td> tags that contains numbers only. They will be replaced with the vnd.ms-excel.numberformat attribute, that will show the leading zeros and is sortable too!

This attribute is tested with Excel 2010, maybe some other users can have a look at it with other versions?