Update GridView.php
Enhancement to let the widget user choose if to pull the toolbar on the left or on the right.

The user can chhose it in two ways:
- when he calls the widget with a toolbar array, he can use the toolbar array attribute 'position':

'toolbar' => [
                     'position' => 'left',
                     'content' => [...],
                     ...
                    ]

- when he calls the widget without a toolbar array (so it uses the default toolbar), he can use the widget attribute 'toolbarPos':

...
'toolbarPos' => 'left'.

The only two values accepted are 'left' and 'right' (as you can see at rows 1803-1809). Any other value will be unconsidered.